+++ src/settingsDialog.hpp
@@ -50,6 +50,7 @@ private:
 
 	// Controller
 	Q_SLOT void on_enableAEMCacheCheckBox_toggled(bool checked);
+	Q_SLOT void on_fullAEMEnumerationCheckBox_toggled(bool checked);
 
 	// Network
 	Q_SLOT void on_protocolComboBox_currentIndexChanged(int index);