+++ src/settingsManager/settings.hpp
@@ -50,6 +50,7 @@ static SettingsManager::SettingDefault Network_InterfaceTypeWiFi = { "avdecc/net
 
 // Controller settings
 static SettingsManager::SettingDefault Controller_AemCacheEnabled = { "avdecc/controller/enableAemCache", false };
+static SettingsManager::SettingDefault Controller_FullStaticModelEnabled = { "avdecc/controller/fullStaticModel", false };
 
 // Settings with no default initial value (no need to register with the SettingsManager) - Not allowed to call registerSettingObserver for those
 static SettingsManager::Setting InterfaceID = { "interfaceID" };